Move homeworld region-selection helpers out of analytics.
Keep the selection store free of analytics imports by owning presets and overlay-derived index helpers in lib.
